move function to own file
